ShaeIt ТЗ 15 pull requests #2-7 postman ok
