+++ src/main/java/ru/practicum/shareit/ShareItApp.java
@@ -2,8 +2,12 @@ package ru.practicum.shareit; // 14
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
+@ComponentScan(basePackages = "ru.practicum.shareit.*")
+@EntityScan("ru.practicum.shareit.*")
 public class ShareItApp {
 
     public static void main(String[] args) {
