--- src/main/java/ru/practicum/shareit/service/EntityCheck.java
@@ -1,9 +1,0 @@
-package ru.practicum.shareit.service;
-
-public interface EntityCheck {
-
-    void isCheckFromSize(int from, int size);
-
-    void isCheckUserId(Long id);
-
-}
