--- src/main/java/ru/practicum/shareit/service/EntityCheckImpl.java
@@ -1,43 +1,0 @@
-package ru.practicum.shareit.service;
-
-import lombok.AllArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-import ru.practicum.shareit.user.model.User;
-import ru.practicum.shareit.user.service.UserService;
-
-@Slf4j
-@Service
-@AllArgsConstructor
-public class EntityCheckImpl implements EntityCheck {
-
-private final UserService userService;
-
-    @Override
-    public void isCheckFromSize(int from, int size) {
-        StackTraceElement[] stackTraceElements = Thread.currentThread().getStackTrace();
-        String messageClass = "";
-        if (stackTraceElements.length >= 3) {
-            StackTraceElement element = stackTraceElements[2];
-            messageClass = element.getClassName() + ":" + element.getMethodName();
-        }
-        log.info("Проверка from {} и size {} вызов из > {} ", from, size, messageClass);
-        if (from < 0 || size < 1) {
-            throw new ru.practicum.shareit.exceptions.EntityNotAvailable("Ошибочный параметр \"size\" или \"from\"");
-        }
-    }
-
-    @Override
-    public void isCheckUserId(Long id) {
-        StackTraceElement[] stackTraceElements = Thread.currentThread().getStackTrace();
-        String messageClass = "";
-        if (stackTraceElements.length >= 3) {
-            StackTraceElement element = stackTraceElements[2];
-            messageClass = element.getClassName() + ":" + element.getMethodName();
-        }
-        log.info("Проверка наличия пользователя с УИН  {} вызов из > {} ", id, messageClass);
-        User user = userService.findUserById(id);
-    }
-
-}
-
