+++ src/main/java/ru/practicum/shareit/user/service/UserService.java
@@ -2,9 +2,11 @@ package ru.practicum.shareit.user.service;
 
 import java.util.List;
 
+import org.springframework.stereotype.Component;
 import ru.practicum.shareit.user.dto.UserDto;
 import ru.practicum.shareit.user.model.User;
 
+@Component
 public interface UserService {
 
     List<UserDto> getUsers();
@@ -19,4 +21,5 @@ public interface UserService {
 
     User findUserById(Long id);
 
+    void isCheckUserId(Long id);
 }
