+++ src/test/java/ru/practicum/shareit/ShareItTests.java
@@ -1,9 +1,12 @@
 package ru.practicum.shareit;
 
 import org.junit.jupiter.api.Test;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootTest
+
 class ShareItTests {
 
 	@Test
