+++ src/test/java/ru/practicum/shareit/booking/BookingControllerTest.java
@@ -14,15 +14,17 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
-import ru.practicum.shareit.service.EntityCheck;
+
 import ru.practicum.shareit.booking.dto.BookingDto;
 import ru.practicum.shareit.booking.dto.BookingModelDto;
 import ru.practicum.shareit.booking.model.Status;
@@ -32,7 +34,7 @@ import ru.practicum.shareit.item.model.Item;
 import ru.practicum.shareit.user.dto.UserDto;
 import ru.practicum.shareit.user.model.User;
 
-@WebMvcTest(controllers = {BookingController.class, EntityCheck.class})
+@WebMvcTest(controllers = BookingController.class)
 @AutoConfigureMockMvc
 class BookingControllerTest {
 
@@ -47,9 +49,6 @@ class BookingControllerTest {
 
     @Autowired
     ObjectMapper mapper;
-
-    @Autowired
-    EntityCheck entityCheck;
 
     User user;
     User owner;
